+++ src/stories/containers/cu-table/cu-table-2.mvvm.ts
@@ -93,7 +93,7 @@ export const useCoreUnitsTableMvvm = () => {
 
   const clearFilters = () => {
     router.push({
-      pathname: '/',
+      pathname: '/core-units',
       search: '',
     });
