+++ tsconfig.json
@@ -16,6 +16,7 @@
     "incremental": true,
     "baseUrl": ".",
     "paths": {
+      "@ses/containers/*": ["src/stories/containers/*"],
       "@ses/components/*": ["src/stories/components/*"],
       "@ses/core/*": ["src/core/*"],
       "@ses/config/*": ["src/config/*"]
